requirements.txt
tensorflow==2.15.0
numpy==1.26.0
Pillow==10.0.1